requirements.txt
streamlit
pandas
openai
python-dotenv
requests
lxml
xmltodict
backoff
pytest
loguru
